requirements.txt
python-telegram-bot==13.7  
requests==2.22.0
python-dotenv==1.0.0 
psycopg2